quick-gradle: simple script to build to .m2
Often, for example in a Dockerfile of a downstream repository,
it's necessary to build all the jars from upstream components
in order to get travis to pass. This script is intended to be
a best effort for "build stuff" which can be placed called by
docker or a user locally.

see: https://github.com/openmicroscopy/openmicroscopy/pull/6001